# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder API
  description: VEuPathDB Strategy Builder Chatbot API
  version: 1.0.0
paths:
...
  /api/v1/strategies:
    get:
      tags:
      - strategies
      summary: List Strategies
      description: List user's saved strategies.
      operationId: list_strategies_api_v1_strategies_get
      security:
      - APIKeyCookie: []
      parameters:
      - name: siteId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategySummaryResponse'
                title: Response List Strategies Api V1 Strategies Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - strategies
      summary: Create Strategy
      description: Create a new strategy.
      operationId: create_strategy_api_v1_strategies_post
      security:
      - APIKeyCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}:
    get:
      tags:
      - strategies
      summary: Get Strategy
      description: Get a strategy by ID.
      operationId: get_strategy_api_v1_strategies__strategyId__get
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - strategies
      summary: Update Strategy
      description: Update a strategy.
      operationId: update_strategy_api_v1_strategies__strategyId__patch
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - strategies
      summary: Delete Strategy
      description: Delete a strategy.
      operationId: delete_strategy_api_v1_strategies__strategyId__delete
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ColocationParams:
      properties:
        upstream:
          type: integer
          minimum: 0.0
          title: Upstream
        downstream:
          type: integer
          minimum: 0.0
          title: Downstream
        strand:
          type: string
          title: Strand
          default: both
      type: object
      required:
      - upstream
      - downstream
      title: ColocationParams
    CreateStrategyRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        siteId:
          type: string
          title: Siteid
        plan:
          $ref: '#/components/schemas/StrategyPlan-Input'
      type: object
      required:
      - name
      - siteId
      - plan
      title: CreateStrategyRequest
      description: Request to create a strategy.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JSONArray:
      items:
        $ref: '#/components/schemas/JSONValue'
      type: array
    JSONObject:
      additionalProperties:
        $ref: '#/components/schemas/JSONValue'
      type: object
    JSONValue: {}
...
    MessageResponse:
      properties:
        role:
          type: string
          title: Role
        content:
          type: string
          title: Content
        toolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Toolcalls
        subKaniActivity:
          anyOf:
          - $ref: '#/components/schemas/SubKaniActivityResponse'
          - type: 'null'
        mode:
          anyOf:
          - type: string
            enum:
            - execute
            - plan
          - type: 'null'
          title: Mode
        citations:
          anyOf:
          - $ref: '#/components/schemas/JSONArray'
          - type: 'null'
        planningArtifacts:
          anyOf:
          - $ref: '#/components/schemas/JSONArray'
          - type: 'null'
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - role
      - content
      - timestamp
      title: MessageResponse
      description: Chat message.
...
    PlanMetadata:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        siteId:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
      type: object
      title: PlanMetadata
    PlanNode-Input:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        displayName:
          anyOf:
          - type: string
          - type: 'null'
          title: Displayname
        filters:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepFilterSpec'
            type: array
          - type: 'null'
          title: Filters
        analyses:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepAnalysisSpec-Input'
            type: array
          - type: 'null'
          title: Analyses
        reports:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepReportSpec-Input'
            type: array
          - type: 'null'
          title: Reports
        searchName:
          type: string
          title: Searchname
        parameters:
          $ref: '#/components/schemas/JSONObject'
        primaryInput:
          anyOf:
          - $ref: '#/components/schemas/PlanNode-Input'
          - type: 'null'
        secondaryInput:
          anyOf:
          - $ref: '#/components/schemas/PlanNode-Input'
          - type: 'null'
        operator:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator
        colocationParams:
          anyOf:
          - $ref: '#/components/schemas/ColocationParams'
          - type: 'null'
      additionalProperties: true
      type: object
      required:
      - searchName
      title: PlanNode
      description: 'Untyped recursive plan node (WDK-aligned).


        Kind is inferred from structure:

        - combine: primaryInput + secondaryInput

        - transform: primaryInput only

        - search: no inputs'
...
    StepAnalysisResponse:
      properties:
        analysisType:
          type: string
          title: Analysistype
        parameters:
          $ref: '#/components/schemas/JSONObject'
        customName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customname
      type: object
      required:
      - analysisType
      title: StepAnalysisResponse
      description: Analysis configuration attached to a step.
...
    StepAnalysisSpec-Input:
      properties:
        analysisType:
          type: string
          title: Analysistype
        parameters:
          $ref: '#/components/schemas/JSONObject'
        customName:
          anyOf:
          - type: string
          - type: 'null'
          title: Customname
      type: object
      required:
      - analysisType
      title: StepAnalysisSpec
...
    StepFilterResponse:
      properties:
        name:
          type: string
          title: Name
        value:
          $ref: '#/components/schemas/JSONValue'
        disabled:
          type: boolean
          title: Disabled
          default: false
      type: object
      required:
      - name
      - value
      title: StepFilterResponse
      description: Filter attached to a step.
    StepFilterSpec:
      properties:
        name:
          type: string
          title: Name
        value:
          $ref: '#/components/schemas/JSONValue'
        disabled:
          type: boolean
          title: Disabled
          default: false
      type: object
      required:
      - name
      - value
      title: StepFilterSpec
...
    StepReportResponse:
      properties:
        reportName:
          type: string
          title: Reportname
          default: standard
        config:
          $ref: '#/components/schemas/JSONObject'
      type: object
      title: StepReportResponse
      description: Report configuration attached to a step.
...
    StepReportSpec-Input:
      properties:
        reportName:
          type: string
          title: Reportname
          default: standard
        config:
          $ref: '#/components/schemas/JSONObject'
      type: object
      title: StepReportSpec
...
    StepResponse:
      properties:
        id:
          type: string
          title: Id
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        displayName:
          type: string
          title: Displayname
        searchName:
          anyOf:
          - type: string
          - type: 'null'
          title: Searchname
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        parameters:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
        operator:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator
        colocationParams:
          anyOf:
          - $ref: '#/components/schemas/JSONObject'
          - type: 'null'
        primaryInputStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Primaryinputstepid
        secondaryInputStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Secondaryinputstepid
        resultCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultcount
        wdkStepId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstepid
        filters:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepFilterResponse'
            type: array
          - type: 'null'
          title: Filters
        analyses:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepAnalysisResponse'
            type: array
          - type: 'null'
          title: Analyses
        reports:
          anyOf:
          - items:
              $ref: '#/components/schemas/StepReportResponse'
            type: array
          - type: 'null'
          title: Reports
      type: object
      required:
      - id
      - displayName
      title: StepResponse
      description: Strategy step.
    StrategyPlan-Input:
      properties:
        recordType:
          type: string
          title: Recordtype
        root:
          $ref: '#/components/schemas/PlanNode-Input'
        metadata:
          anyOf:
          - $ref: '#/components/schemas/PlanMetadata'
          - type: 'null'
      additionalProperties: true
      type: object
      required:
      - recordType
      - root
      title: StrategyPlan
...
    StrategyResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        siteId:
          type: string
          title: Siteid
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        steps:
          items:
            $ref: '#/components/schemas/StepResponse'
          type: array
          title: Steps
        rootStepId:
          anyOf:
          - type: string
          - type: 'null'
          title: Rootstepid
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        messages:
          anyOf:
          - items:
              $ref: '#/components/schemas/MessageResponse'
            type: array
          - type: 'null'
          title: Messages
        thinking:
          anyOf:
          - $ref: '#/components/schemas/ThinkingResponse'
          - type: 'null'
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - name
      - siteId
      - recordType
      - steps
      - rootStepId
      - createdAt
      - updatedAt
      title: StrategyResponse
      description: Full strategy with steps.
    StrategySummaryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        siteId:
          type: string
          title: Siteid
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        stepCount:
          type: integer
          title: Stepcount
        resultCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultcount
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - name
      - siteId
      - recordType
      - stepCount
      - createdAt
      - updatedAt
      title: StrategySummaryResponse
      description: Strategy summary for list views.
    SubKaniActivityResponse:
      properties:
        calls:
          additionalProperties:
            items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          type: object
          title: Calls
        status:
          additionalProperties:
            type: string
          type: object
          title: Status
      type: object
      required:
      - calls
      - status
      title: SubKaniActivityResponse
      description: Sub-kani tool call activity.
    ThinkingResponse:
      properties:
        toolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Toolcalls
        lastToolCalls:
          anyOf:
          - items:
              $ref: '#/components/schemas/ToolCallResponse'
            type: array
          - type: 'null'
          title: Lasttoolcalls
        subKaniCalls:
          anyOf:
          - additionalProperties:
              items:
                $ref: '#/components/schemas/ToolCallResponse'
              type: array
            type: object
          - type: 'null'
          title: Subkanicalls
        subKaniStatus:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Subkanistatus
        reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning
        updatedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updatedat
      type: object
      title: ThinkingResponse
      description: In-progress tool call state.
...
    ToolCallResponse:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        arguments:
          $ref: '#/components/schemas/JSONObject'
        result:
          anyOf:
          - type: string
          - type: 'null'
          title: Result
      type: object
      required:
      - id
      - name
      - arguments
      title: ToolCallResponse
      description: Tool call information.
...
    UpdateStrategyRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        plan:
          anyOf:
          - $ref: '#/components/schemas/StrategyPlan-Input'
          - type: 'null'
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
      type: object
      title: UpdateStrategyRequest
      description: Request to update a strategy.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: pathfinder-auth
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.pathfinder.veupathdb.org
  description: Production